Move clientForResource into utils
