+++ cmd/delete.go
@@ -87,7 +87,7 @@ var deleteCmd = &cobra.Command{
 			desc := fmt.Sprintf("%s/%s", obj.GetKind(), fqName(obj))
 			log.Info("Deleting ", desc)
 
-			c, err := clientForResource(clientpool, disco, obj, defaultNs)
+			c, err := utils.ClientForResource(clientpool, disco, obj, defaultNs)
 			if err != nil {
 				return err
 			}
