+++ utils/client.go
@@ -16,13 +16,18 @@
 package utils
 
 import (
+	"fmt"
 	"sync"
 
 	"github.com/emicklei/go-restful/swagger"
+	log "github.com/sirupsen/logrus"
+	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/version"
 	"k8s.io/client-go/discovery"
+	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/rest"
 )
 
@@ -102,3 +107,47 @@ func (c *memcachedDiscoveryClient) SwaggerSchema(version schema.GroupVersion) (*
 }
 
 var _ discovery.CachedDiscoveryInterface = &memcachedDiscoveryClient{}
+
+// ClientForResource returns the ResourceClient for a given object
+func ClientForResource(pool dynamic.ClientPool, disco discovery.DiscoveryInterface, obj runtime.Object, defNs string) (*dynamic.ResourceClient, error) {
+	gvk := obj.GetObjectKind().GroupVersionKind()
+
+	client, err := pool.ClientForGroupVersionKind(gvk)
+	if err != nil {
+		return nil, err
+	}
+
+	resource, err := serverResourceForGroupVersionKind(disco, gvk)
+	if err != nil {
+		return nil, err
+	}
+
+	meta, err := meta.Accessor(obj)
+	if err != nil {
+		return nil, err
+	}
+	namespace := meta.GetNamespace()
+	if namespace == "" {
+		namespace = defNs
+	}
+
+	log.Debugf("Fetching client for %s namespace=%s", resource, namespace)
+	rc := client.Resource(resource, namespace)
+	return rc, nil
+}
+
+func serverResourceForGroupVersionKind(disco discovery.DiscoveryInterface, gvk schema.GroupVersionKind) (*metav1.APIResource, error) {
+	resources, err := disco.ServerResourcesForGroupVersion(gvk.GroupVersion().String())
+	if err != nil {
+		return nil, err
+	}
+
+	for _, r := range resources.APIResources {
+		if r.Kind == gvk.Kind {
+			log.Debugf("Chose API '%s' for %s", r.Name, gvk)
+			return &r, nil
+		}
+	}
+
+	return nil, fmt.Errorf("Server is unable to handle %s", gvk)
+}
